Use multipage layout
